network-license-manager/tests/requirements.txt
docker
pytest-testinfra